+++ product_code/examreg_service/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Refactor
 ## Bugfix
  - Fixed missing behaviour for "GetAllExamregsHyperledger" command
+ - Fixed sequence serialization error through a wrapper class
 
 # [v0.15.0](https://github.com/upb-uc4/University-Credits-4.0/compare/examreg-v0.14.1...examreg-v0.15.0) (2020-12-20)
 ## Feature